+++ lib/cjs.ts
@@ -13,10 +13,9 @@ import cjsPlatformShim from './platform-shims/cjs.js';
 
 // See https://github.com/yargs/yargs#supported-nodejs-versions for our
 // version support policy. The YARGS_MIN_NODE_VERSION is used for testing only.
-const minNodeVersion =
-  process && process.env && process.env.YARGS_MIN_NODE_VERSION
-    ? Number(process.env.YARGS_MIN_NODE_VERSION)
-    : 10;
+const minNodeVersion = process?.env?.YARGS_MIN_NODE_VERSION
+  ? Number(process.env.YARGS_MIN_NODE_VERSION)
+  : 12;
 if (process && process.version) {
   const major = Number(process.version.match(/v([^.]+)/)![1]);
   if (major < minNodeVersion) {