+++ lib/validation.ts
@@ -154,7 +154,7 @@ export function validation(
 
     Object.keys(argv).forEach(key => {
       if (
-        specialKeys.indexOf(key) === -1 &&
+        !specialKeys.includes(key) &&
         !Object.prototype.hasOwnProperty.call(positionalMap, key) &&
         !Object.prototype.hasOwnProperty.call(
           yargs.getInternalMethods().getParseContext(),
@@ -173,13 +173,33 @@ export function validation(
         isDefaultCommand)
     ) {
       argv._.slice(currentContext.commands.length).forEach(key => {
-        if (commandKeys.indexOf('' + key) === -1) {
+        if (!commandKeys.includes('' + key)) {
           unknown.push('' + key);
         }
       });
     }
 
-    if (unknown.length > 0) {
+    // https://github.com/yargs/yargs/issues/1861
+    if (checkPositionals) {
+      // Check for non-option args that are not in currentContext.commands
+      // Take into account expected args from commands and yargs.demand(number)
+      const demandedCommands = yargs.getDemandedCommands();
+      const maxNonOptDemanded = demandedCommands._?.max || 0;
+      const expected = currentContext.commands.length + maxNonOptDemanded;
+      if (expected < argv._.length) {
+        argv._.slice(expected).forEach(key => {
+          key = String(key);
+          if (
+            !currentContext.commands.includes(key) &&
+            !unknown.includes(key)
+          ) {
+            unknown.push(key);
+          }
+        });
+      }
+    }
+
+    if (unknown.length) {
       usage.fail(
         __n(
           'Unknown argument: %s',
@@ -201,7 +221,7 @@ export function validation(
 
     if (currentContext.commands.length > 0 || commandKeys.length > 0) {
       argv._.slice(currentContext.commands.length).forEach(key => {
-        if (commandKeys.indexOf('' + key) === -1) {
+        if (!commandKeys.includes('' + key)) {
           unknown.push('' + key);
         }
       });
@@ -232,15 +252,10 @@ export function validation(
       return false;
     }
     const newAliases = (yargs.parsed as DetailedArguments).newAliases;
-    for (const a of [key, ...aliases[key]]) {
-      if (
-        !Object.prototype.hasOwnProperty.call(newAliases, a) ||
-        !newAliases[key]
-      ) {
-        return true;
-      }
-    }
-    return false;
+    return [key, ...aliases[key]].some(
+      a =>
+        !Object.prototype.hasOwnProperty.call(newAliases, a) || !newAliases[key]
+    );
   };
 
   // validate arguments limited to enumerated choices
@@ -324,10 +339,10 @@ export function validation(
     } else if (val.match(/^--no-.+/)) {
       // check if key/value doesn't exist
       val = val.match(/^--no-(.+)/)[1];
-      val = !argv[val];
+      val = !Object.prototype.hasOwnProperty.call(argv, val);
     } else {
       // check if key/value exists
-      val = argv[val];
+      val = Object.prototype.hasOwnProperty.call(argv, val);
     }
     return val;
   }