+++ lib/yargs-factory.ts
@@ -547,9 +547,8 @@ export class YargsInstance {
     if (typeof value === 'function') {
       assertSingleKey(key, this.#shim);
       if (!this.#options.defaultDescription[key])
-        this.#options.defaultDescription[key] = this.#usage.functionDescription(
-          value
-        );
+        this.#options.defaultDescription[key] =
+          this.#usage.functionDescription(value);
       value = value.call();
     }
     this[kPopulateParserHintSingleValueDictionary]<'default'>(
@@ -810,7 +809,7 @@ export class YargsInstance {
       );
     } else {
       globals.forEach(g => {
-        if (this.#options.local.indexOf(g) === -1) this.#options.local.push(g);
+        if (!this.#options.local.includes(g)) this.#options.local.push(g);
       });
     }
     return this;
@@ -1032,11 +1031,7 @@ export class YargsInstance {
     );
     this[kFreeze](); // Push current state of parser onto stack.
     if (typeof args === 'undefined') {
-      const argv = this[kRunYargsParserAndExecuteCommands](this.#processArgs);
-      const tmpParsed = this.parsed;
-      this[kUnfreeze](); // Pop the stack.
-      this.parsed = tmpParsed;
-      return argv;
+      args = this.#processArgs;
     }
 
     // a context object can optionally be provided, this allows
@@ -1063,6 +1058,7 @@ export class YargsInstance {
       args,
       !!shortCircuit
     );
+    const tmpParsed = this.parsed;
     this.#completion!.setParsed(this.parsed as DetailedArguments);
     if (isPromise(parsed)) {
       return parsed
@@ -1082,10 +1078,12 @@ export class YargsInstance {
         })
         .finally(() => {
           this[kUnfreeze](); // Pop the stack.
+          this.parsed = tmpParsed;
         });
     } else {
       if (this.#parseFn) this.#parseFn(this.#exitError, parsed, this.#output);
       this[kUnfreeze](); // Pop the stack.
+      this.parsed = tmpParsed;
     }
     return parsed;
   }
@@ -1095,11 +1093,9 @@ export class YargsInstance {
     _parseFn?: ParseCallback
   ): Promise<Arguments> {
     const maybePromise = this.parse(args, shortCircuit, _parseFn);
-    if (!isPromise(maybePromise)) {
-      return Promise.resolve(maybePromise);
-    } else {
-      return maybePromise;
-    }
+    return !isPromise(maybePromise)
+      ? Promise.resolve(maybePromise)
+      : maybePromise;
   }
   parseSync(
     args?: string | string[],
@@ -1160,17 +1156,15 @@ export class YargsInstance {
       'alias',
     ];
     opts = objFilter(opts, (k, v) => {
-      let accept = supportedOpts.indexOf(k) !== -1;
       // type can be one of string|number|boolean.
-      if (k === 'type' && ['string', 'number', 'boolean'].indexOf(v) === -1)
-        accept = false;
-      return accept;
+      if (k === 'type' && !['string', 'number', 'boolean'].includes(v))
+        return false;
+      return supportedOpts.includes(k);
     });
 
     // copy over any settings that can be inferred from the command string.
-    const fullCommand = this.#context.fullCommands[
-      this.#context.fullCommands.length - 1
-    ];
+    const fullCommand =
+      this.#context.fullCommands[this.#context.fullCommands.length - 1];
     const parseOptions = fullCommand
       ? this.#command.cmdToParseOptions(fullCommand)
       : {
@@ -1447,7 +1441,7 @@ export class YargsInstance {
         return;
       const hint = this.#options[hintKey];
       if (Array.isArray(hint)) {
-        if (~hint.indexOf(optionKey)) hint.splice(hint.indexOf(optionKey), 1);
+        if (hint.includes(optionKey)) hint.splice(hint.indexOf(optionKey), 1);
       } else if (typeof hint === 'object') {
         delete (hint as Dictionary)[optionKey];
       }
@@ -1728,9 +1722,8 @@ export class YargsInstance {
       postProcess: this[kPostProcess].bind(this),
       reset: this[kReset].bind(this),
       runValidation: this[kRunValidation].bind(this),
-      runYargsParserAndExecuteCommands: this[
-        kRunYargsParserAndExecuteCommands
-      ].bind(this),
+      runYargsParserAndExecuteCommands:
+        this[kRunYargsParserAndExecuteCommands].bind(this),
       setHasOutput: this[kSetHasOutput].bind(this),
     };
   }
@@ -1790,10 +1783,8 @@ export class YargsInstance {
   [kReset](aliases: Aliases = {}): YargsInstance {
     this.#options = this.#options || ({} as Options);
     const tmpOptions = {} as Options;
-    tmpOptions.local = this.#options.local ? this.#options.local : [];
-    tmpOptions.configObjects = this.#options.configObjects
-      ? this.#options.configObjects
-      : [];
+    tmpOptions.local = this.#options.local || [];
+    tmpOptions.configObjects = this.#options.configObjects || [];
 
     // if a key has been explicitly set as local,
     // we should reset it before passing options to command.
@@ -1974,7 +1965,7 @@ export class YargsInstance {
           .concat(aliases[this.#helpOpt] || [])
           .filter(k => k.length > 1);
         // check if help should trigger and strip it from _.
-        if (~helpCmds.indexOf('' + argv._[argv._.length - 1])) {
+        if (helpCmds.includes('' + argv._[argv._.length - 1])) {
           argv._.pop();
           helpOptSet = true;
         }
@@ -1983,13 +1974,12 @@ export class YargsInstance {
       const handlerKeys = this.#command.getCommands();
       const requestCompletions = this.#completion!.completionKey in argv;
       const skipRecommendation = helpOptSet || requestCompletions || helpOnly;
-
       if (argv._.length) {
         if (handlerKeys.length) {
           let firstUnknownCommand;
           for (let i = commandIndex || 0, cmd; argv._[i] !== undefined; i++) {
             cmd = String(argv._[i]);
-            if (~handlerKeys.indexOf(cmd) && cmd !== this.#completionCommand) {
+            if (handlerKeys.includes(cmd) && cmd !== this.#completionCommand) {
               // commands are executed using a recursive algorithm that executes
               // the deepest command first; we keep track of the position in the
               // argv._ array that is currently being executed.
@@ -2035,7 +2025,7 @@ export class YargsInstance {
         // generate a completion script for adding to ~/.bashrc.
         if (
           this.#completionCommand &&
-          ~argv._.indexOf(this.#completionCommand) &&
+          argv._.includes(this.#completionCommand) &&
           !requestCompletions
         ) {
           if (this.#exitProcess) setBlocking(true);
@@ -2111,7 +2101,7 @@ export class YargsInstance {
         );
       }
 
-      // If the help or version options where used and exitProcess is false,
+      // If the help or version options were used and exitProcess is false,
       // or if explicitly skipped, we won't run validations.
       if (!skipValidation) {
         if (parsed.error) throw new YError(parsed.error.message);
@@ -2159,8 +2149,6 @@ export class YargsInstance {
     parseErrors: Error | null,
     isDefaultCommand?: boolean
   ): (argv: Arguments) => void {
-    aliases = {...aliases};
-    positionalMap = {...positionalMap};
     const demandedOptions = {...this.getDemandedOptions()};
     return (argv: Arguments) => {
       if (parseErrors) throw new YError(parseErrors.message);